Extract handling line item property change
